convert ot np.uint8

+++ apps/siamese/siamese_app.py
@@ -94,7 +94,7 @@ class SiameseApp(object):
             dv = distance.cpu().data.numpy()[0]
             img0_raw = x0.cpu().numpy().reshape((3, 224, 224))
             print('x0_a: {0};'.format(img0_raw.shape))
-            img0 = Image.fromarray(img0_raw)
+            img0 = Image.fromarray(np.uint8(img0_raw))
             img0.save('./logs/img0_{0}_{1}.jpg'.format(i, dv))
             img1 = Image.fromarray(x1.cpu().numpy())
             img1.save('./logs/img1_{0}_{1}.jpg'.format(x1, dv))
